Increased AI Upload Security
:cl:
add: You now require an ID with AI Upload access to use the Borg and AI upload consoles.
add: Emags can break the security protocols on these consoles, allowing them to be usable by anyone but they will require 60 seconds to do so, and this will alert Security over the security channel with the location of the hacked console.
/:cl: