Added Speed Upgrade and added it to the vendor, tweaked vendor prices for lava-proofing and added a new multiplicative_slowdown for minebot_speedupgrade
The multiplicative_slowdown of the Minebot is reduced from 3 to 2 with this upgrade, making it move 33% faster. The price for it (and now for the lava-proofing upgrade too) is 750 points. Also, added minor chat feedbacks for all the upgrades, so that players know that they successfully added upgrades to their Minebots, while also changing existing ones because [src] was returning the name of the upgrade and not the nameo f the Minebot. Also made sure to make it obvious when I made fulp-exclusive changes.