Removed \improper in name and added serial number to name, minebots now also drop their ores on death.
Minebots weren't dropping ores when they died and it was an obvious typoe, since 0 and ) are on the same key. I also expanded on my comment about the Minebot GPS signal, because I hope someone will help me figuring that one out. Name change will make it much easier for admins to know which minebot did what when they decide to grief as minebots or whatever.