Added round_end_antags.json and added .json as a valid extension for files viewable by people
Hopefully this won't cause any issue, otherwise I can just change round_end_antags and round_end_data to be .txt files, I tested it and it works fine as well. Just did this because folks on coderbus told me that it was the way to do it. Now, round_end_data.json is also viewable, even if I doubt anyone would want to look at that useless mess.