This time, ACTUALLY makes it fulp-only. Promise.
It was lazily done last time because I was extremely tired. This ACTUALLY makes it a Fulp-only martial art, so entirely seperate of the main Wrestling in case Admins want it for whatever reason. I also realized I forgot one of the moves still putting you to sleep, so I removed that too.