Update lavaland_surface_ash_walker1.dmm
Did a major overhaul of the Necropolis to make it into a real village for the ashwalkers. Moved the temple to the middle of town, added a chief's home to make rulings from with some trophy items, a general longhouse with 4 hunter/guard chests, 1 architect chest, 1 farmer chest. Reworked the Necropolis farm. Added seeds required to make Aloe Cream and Mourning Poultice. Added a shaman's hut with makeshift ashwalker surgery kit and mortar and pestle, added a general store for interaction with other ashwalkers/visitors. Gave the entire area storm coverage.